refactor: move code to separate files
