+++ packages/echo-server/package.json
@@ -5,7 +5,6 @@
   "author": "Scalar (https://github.com/scalar)",
   "bugs": "https://github.com/scalar/scalar/issues/new",
   "dependencies": {
-    "body-parser": "1.20.2",
     "cookie-parser": "1.4.6",
     "cors": "2.8.5",
     "express": "4.18.2"
