+++ packages/echo-server/src/routes/oauth2/db.ts
@@ -1,0 +1,23 @@
+export const db = {
+  // Here is a fast overview of what your db model should look like
+  authorizationCode: {
+    authorizationCode: '', // A string that contains the code
+    expiresAt: new Date(), // A date when the code expires
+    redirectUri: '', // A string of where to redirect to with this code
+    client: null, // See the client section
+    user: null, // Whatever you want... This is where you can be flexible with the protocol
+  },
+  client: {
+    // Application wanting to authenticate with this server
+    clientId: '', // Unique string representing the client
+    clientSecret: '', // Secret of the client; Can be null
+    grants: [], // Array of grants that the client can use (ie, `authorization_code`)
+    redirectUris: [], // Array of urls the client is allowed to redirect to
+  },
+  token: {
+    accessToken: '', // Access token that the server created
+    accessTokenExpiresAt: new Date(), // Date the token expires
+    client: null, // Client associated with this token
+    user: null, // User associated with this token
+  },
+}
